requirements.txt
# Django Core
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.0,<5.0
django-filter>=23.0,<24.0

# Database
psycopg2-binary>=2.9,<3.0

# Caching & Task Queue
redis>=4.5,<5.0
django-redis>=5.3,<6.0
celery>=5.3,<6.0

# Image Processing
Pillow>=10.0,<11.0

# AWS S3
boto3>=1.28,<2.0

# HTTP Requests
requests>=2.31,<3.0

# Authentication & Security
PyJWT>=2.8,<3.0

# Utilities
python-slugify>=8.0,<9.0
python-dateutil>=2.8,<3.0

# Testing
pytest>=7.4,<8.0
pytest-django>=4.5,<5.0
pytest-cov>=4.1,<5.0

# Code Quality
flake8>=6.0,<7.0
black>=23.0,<24.0
isort>=5.12,<6.0

# Monitoring
sentry-sdk>=1.32,<2.0
psutil>=5.9,<6.0
django-celery-beat>=2.5,<3.0

# Environment
django-environ>=0.11,<1.0

# API Documentation
drf-spectacular>=0.27,<1.0
